requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
django-filter==22.1
django-sequences==2.7
djangorestframework==3.13.1
importlib-metadata==4.12.0
Markdown==3.3.7
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
zipp==3.8.0
